+++ skytemple_files/graphics/chara_wan/model.py
@@ -59,6 +59,8 @@ DIM_TABLE = [(1, 1), (2, 2), (4, 4), (8, 8), \
              (2, 1), (4, 1), (4, 2), (8, 4), \
              (1, 2), (1, 4), (2, 4), (4, 8)]
 
+DEBUG_PRINT = False
+
 
 class WanFile():
